+++ Casks/tokenglance.rb
@@ -1,6 +1,6 @@
 cask "tokenglance" do
   version "0.1.1"
-  sha256 "698730c512c1b04f4193cfe79f8cae2d28ebde75743791bb7f8481ef2e6534ce"
+  sha256 "398d7af0b59b233066774680b89f4dc16c1ca0f704c31b08b352399050f7b145"
 
   url "https://github.com/marcel-breuer/tokenglance/releases/download/v#{version}/TokenGlance-#{version}-arm64.zip"
   name "TokenGlance"